requirements.txt
Pillow>=9.0
imagehash>=4.3
boto3>=1.26
openpyxl>=3.1
psycopg2-binary>=2.9
sshtunnel>=0.4
